Revert "Modify the Dashboard with the new indicator;Optimized the Topology Panel (#183)"
This reverts commit 183472f52d5104a89f9e8b31935ca46a5f5db240.
