+++ grafana-plugins/dashboard-json/topology.json
@@ -28,7 +28,7 @@
         "overrides": []
       },
       "gridPos": {
-        "h": 23,
+        "h": 21,
         "w": 24,
         "x": 0,
         "y": 0
@@ -39,14 +39,14 @@
       "pluginVersion": "7.1.1",
       "targets": [
         {
-          "expr": "increase(kindling_topology_request_total[$__range])",
+          "expr": "increase(kindling_topology_request_duration_nanoseconds_count[$__range])",
           "instant": true,
           "interval": "",
           "legendFormat": "",
           "refId": "A"
         },
         {
-          "expr": "increase(kindling_topology_request_duration_nanoseconds_total[$__range])",
+          "expr": "increase(kindling_topology_request_duration_nanoseconds_sum[$__range])",
           "instant": true,
           "interval": "",
           "legendFormat": "",
@@ -69,18 +69,26 @@
           "refId": "C"
         },
         {
-          "expr": "increase(kindling_entity_request_total[$__range])",
+          "expr": "increase(kindling_entity_request_duration_nanoseconds_count[$__range])",
           "instant": true,
           "interval": "",
           "legendFormat": "",
           "refId": "D"
         },
         {
-          "expr": "sum (increase(kindling_entity_request_duration_nanoseconds_total[$__range])) by (namespace, workload_name, pod)",
+          "expr": "sum (increase(kindling_entity_request_duration_nanoseconds_sum[$__range])) by (namespace, workload_name, pod)",
           "instant": true,
           "interval": "",
           "legendFormat": "",
           "refId": "E"
+        },
+        {
+          "expr": "sum by(namespace, workload_name, pod) (kindling_entity_request_duration_nanoseconds_count{namespace=~\"$namespace\",workload_name=~\"$workload\", response_content=~\"400|404|500|503\"}) / sum by(namespace, workload_name, pod) (kindling_entity_request_duration_nanoseconds_count{namespace=~\"$namespace\",workload_name=~\"$workload\"}) * 100",
+          "hide": true,
+          "instant": true,
+          "interval": "",
+          "legendFormat": "",
+          "refId": "F"
         },
         {
           "expr": "sum(increase(kindling_entity_request_send_bytes_total[$__range])) by(namespace, workload_name, pod)",
@@ -104,7 +112,7 @@
           "refId": "J"
         },
         {
-          "expr": "avg_over_time(kindling_tcp_srtt_microseconds[$__range])",
+          "expr": "avg_over_time(kindling_tcp_rtt_microseconds[$__range])",
           "instant": true,
           "interval": "",
           "legendFormat": "",
@@ -128,10 +136,63 @@
   "style": "dark",
   "tags": [],
   "templating": {
-    "list": []
+    "list": [
+      {
+        "allValue": null,
+        "current": {
+          "selected": false,
+          "text": "All",
+          "value": "$__all"
+        },
+        "datasource": "Prometheus",
+        "definition": "label_values(kindling_entity_request_duration_nanoseconds_count, namespace)",
+        "hide": 0,
+        "includeAll": true,
+        "label": "Namespace",
+        "multi": false,
+        "name": "namespace",
+        "options": [],
+        "query": "label_values(kindling_entity_request_duration_nanoseconds_count, namespace)",
+        "refresh": 2,
+        "regex": "",
+        "skipUrlSync": false,
+        "sort": 2,
+        "tagValuesQuery": "",
+        "tags": [],
+        "tagsQuery": "",
+        "type": "query",
+        "useTags": false
+      },
+      {
+        "allValue": null,
+        "current": {
+          "selected": false,
+          "text": "All",
+          "value": "$__all"
+        },
+        "datasource": "Prometheus",
+        "definition": "kindling_entity_request_duration_nanoseconds_count{namespace =\"$namespace\"}",
+        "hide": 0,
+        "includeAll": true,
+        "label": "Workload",
+        "multi": false,
+        "name": "workload",
+        "options": [],
+        "query": "kindling_entity_request_duration_nanoseconds_count{namespace =\"$namespace\"}",
+        "refresh": 2,
+        "regex": ".*workload_name=\"(.*?)\".*",
+        "skipUrlSync": false,
+        "sort": 0,
+        "tagValuesQuery": "",
+        "tags": [],
+        "tagsQuery": "",
+        "type": "query",
+        "useTags": false
+      }
+    ]
   },
   "time": {
-    "from": "now-30m",
+    "from": "now-2d",
     "to": "now"
   },
   "timepicker": {
